+++ test/decode/language/ruby/.fixtures/block_argument.rb
@@ -1,0 +1,2 @@
+
+define_method(:foo, &block)