+++ releases.md
@@ -1,5 +1,9 @@
 # Releases
 
+## Unreleased
+
+  - Fix handling of `&block` arguments in call nodes.
+
 ## v0.23.4
 
   - Fix handling of definitions nested within `if`/`unless`/`elsif`/`else` blocks.